refactor: extract common function into useColor hook
