+++ packages/components/ripple/index.ts
@@ -3,3 +3,5 @@ import Ripple from './src/index.vue'
 
 export const FnRipple = withInstall(Ripple, 'FnRipple')
 export default FnRipple
+export * from './src/ripple'
+export type RippleInstance = InstanceType<typeof Ripple>
