+++ packages/components/ripple/src/index.vue
@@ -1,6 +1,5 @@
 <script lang="ts" setup>
 import {
-  computed,
   getCurrentInstance,
   onMounted,
   onUnmounted,
@@ -9,7 +8,7 @@ import {
   watchEffect,
 } from 'vue'
 import { addUnit } from '@fusion-ui/utils/dom'
-import useTheme from '@fusion-ui/theme'
+import { useRipple } from '@fusion-ui/hooks'
 import type { RippleStyle } from './ripple'
 import { rippleProps } from './ripple'
 
@@ -17,9 +16,9 @@ const props = defineProps(rippleProps)
 const ripplesArr = reactive<RippleStyle[]>([])
 const duration = ref<number>(600)
 const parent = getCurrentInstance()?.parent
+const { classes } = useRipple(props)
 let bounce: NodeJS.Timeout | null = null
 let listener: EventListener | null = null
-const theme = useTheme()
 
 const addRipple = (event: MouseEvent) => {
   const target = event.currentTarget as HTMLElement
@@ -50,20 +49,6 @@ watchEffect(() => {
   }
 })
 
-const backgroundColor = computed(() => {
-  const { color } = props
-  if (!color) {
-    return null
-  }
-  if (typeof color === 'function') {
-    return color(theme)
-  }
-  if (['primary', 'secondary', 'tertiary', 'error'].includes(color)) {
-    return `var(--md-sys-color-${color})`
-  }
-  return color
-})
-
 onMounted(() => {
   if (parent)
     listener = parent.proxy?.$el.addEventListener('mousedown', addRipple)
@@ -90,12 +75,12 @@ export default {
       v-for="(ripple, index) of ripplesArr"
       :key="`ripple_${index}`"
       :style="{
-        backgroundColor: backgroundColor ?? 'var(--md-sys-color-primary)',
         top: addUnit(ripple.y),
         left: addUnit(ripple.x),
         width: addUnit(ripple.size),
         height: addUnit(ripple.size),
       }"
+      :class="classes['fn-ripple']"
       class="fn-ripple pressed-state-layer"
     />
   </span>
