+++ packages/hooks/index.ts
@@ -7,3 +7,4 @@ export { UseBadge } from './use-badge'
 export { UseProgress } from './use-progress'
 export { UseTag } from './use-tag'
 export { useCheckbox } from './use-checkbox'
+export { useRipple } from './use-ripple'
