+++ packages/theme/package.json
@@ -5,6 +5,9 @@
   "license": "MIT",
   "main": "index.ts",
   "module": "index.ts",
+  "peerDependencies": {
+    "vue": "~3.2.45"
+  },
   "dependencies": {
     "@material/material-color-utilities": "^0.2.7"
   }
